#261 Разнесение генерации и визуализации сгенерированного мира
