+++ Zilon.Core/Zilon.Core/WorldGeneration/Generator.cs
@@ -2,9 +2,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Imaging;
-using System.Linq;
-using System.Runtime.InteropServices;
 
 using Zilon.Core.CommonServices.Dices;
 using Zilon.Core.WorldGeneration.AgentCards;
@@ -20,7 +17,7 @@ namespace Zilon.Core.WorldGeneration
             _dice = dice;
         }
 
-        public void Generate()
+        public Globe Generate()
         {
             const int Size = 20;
             const int StartRealmCount = 4;
@@ -51,8 +48,7 @@ namespace Zilon.Core.WorldGeneration
                 {
                     globe.Terrain[i][j] = new TerrainCell
                     {
-                        X = i,
-                        Y = j
+                        Coords = new OffsetCoords(i, j)
                     };
 
                     scanResult.Free.Add(globe.Terrain[i][j]);
@@ -137,86 +133,9 @@ namespace Zilon.Core.WorldGeneration
             agentsClock.Stop();
             Console.WriteLine(agentsClock.ElapsedMilliseconds / 1f + "s");
 
-            var saveBmpClock = new Stopwatch();
-            saveBmpClock.Start();
-
-            var branchColors = new[] { Color.Red, Color.Blue, Color.Green, Color.Yellow,
-                Color.Black, Color.Magenta, Color.Maroon, Color.LightGray };
-            using (var realmBmp = new DirectBitmap(Size, Size))
-            using (var branchmBmp = new DirectBitmap(Size, Size))
-            {
-                for (var i = 0; i < Size; i++)
-                {
-                    for (var j = 0; j < Size; j++)
-                    {
-                        var cell = globe.Terrain[i][j];
-                        if (globe.localitiesCells.TryGetValue(cell, out var locality))
-                        {
-                            var branch = locality.Branches.Single(x => x.Value > 0);
-
-                            var owner = locality.Owner;
-                            realmBmp.SetPixel(i, j, owner.Color);
-                            branchmBmp.SetPixel(i, j, branchColors[(int)branch.Key]);
-                        }
-                        else
-                        {
-                            realmBmp.SetPixel(i, j, Color.White);
-                            branchmBmp.SetPixel(i, j, Color.White);
-                        }
-                    }
-                }
-
-                realmBmp.Bitmap.Save(@"c:\worldgen\realms.bmp", ImageFormat.Bmp);
-                branchmBmp.Bitmap.Save(@"c:\worldgen\branches.bmp", ImageFormat.Bmp);
-            }
-
-            saveBmpClock.Stop();
-            Console.WriteLine(saveBmpClock.ElapsedMilliseconds / 1f + "s");
+            return globe;
         }
 
-        public class DirectBitmap : IDisposable
-        {
-            public Bitmap Bitmap { get; private set; }
-            public Int32[] Bits { get; private set; }
-            public bool Disposed { get; private set; }
-            public int Height { get; private set; }
-            public int Width { get; private set; }
-
-            protected GCHandle BitsHandle { get; private set; }
-
-            public DirectBitmap(int width, int height)
-            {
-                Width = width;
-                Height = height;
-                Bits = new Int32[width * height];
-                BitsHandle = GCHandle.Alloc(Bits, GCHandleType.Pinned);
-                Bitmap = new Bitmap(width, height, width * 4, PixelFormat.Format32bppPArgb, BitsHandle.AddrOfPinnedObject());
-            }
-
-            public void SetPixel(int x, int y, Color colour)
-            {
-                int index = x + (y * Width);
-                int col = colour.ToArgb();
-
-                Bits[index] = col;
-            }
-
-            public Color GetPixel(int x, int y)
-            {
-                int index = x + (y * Width);
-                int col = Bits[index];
-                Color result = Color.FromArgb(col);
-
-                return result;
-            }
-
-            public void Dispose()
-            {
-                if (Disposed) return;
-                Disposed = true;
-                Bitmap.Dispose();
-                BitsHandle.Free();
-            }
-        }
+        
     }
 }
